Renamed "Group10" to "XPlaneReflection"
