+++ XPlaneReflection/XPlaneReflection.c
@@ -1,4 +1,4 @@
-#include "Group10.h"
+#include "XPlaneReflection.h"
 
 void xPlaneReflection() {
 }
